[2021, day 24, part 1 & 2] Reverse engineer assembly program using z3
We are given a blob of assembly and the semantics of the language, and
have to find inputs for the programs: being the lexicographically
largest / smallest integer accepted by the program.

The program turns out to be 14 repetitions of the same structure, with
different numerical parameters (some kind of polynomial), which allows
to reduce the complexity of the description of the input.

Then, the reifed blob is fed into Z3, along with some constraints.