[2021, day 18, part 1 & 2] Tree-like data structure, reduction rules
This problem was conceptually on the harder side. The gist of my
solution is to define a tree-like data structure to represent "numbers",
and apply the reduction rules in place, keeping track of progress. The
only difficulty of part 2 stands from the fact that there it's easy to
accidentally reuse a part of a number that has been modified through
reduction, during part 1. Therefore, I perform a deep-copy of the list
of numbers before going through part 1, and then, make sure to have a
proper copy of each number for part 2.