[2021, day 19, part 1 & 2] 3D geometry, scan matching
This problem was the hardest so far for me. I initially tried to rely on
heuristic to deal with the matching of scanners (identify points in a
unique manner, somehow), but this turned out to be atrociously slow. The
final solution simply iterate through all possible rotations and offset
combinations until finding something that matches.